requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# API and web framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# CLI
click>=8.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utilities
python-dateutil>=2.8.0
requests>=2.26.0
tqdm>=4.62.0

# Development
black>=21.7b0
flake8>=3.9.0
isort>=5.9.0 